+++ packages/react-sprucebot/src/components/Calendar/BigCalendar.js
@@ -38,7 +38,7 @@ export default class BigCalendar extends Component {
 			renderAllEvents: true,
 			showAllTeammates: props.defaultMode === 'team',
 			transitioning: false,
-			selectedDate: moment(),
+			selectedDate: moment().tz(props.auth.Location.timezone),
 			earliestTime: null,
 			latestTime: null,
 			teammates: props.teammates ? props.teammates : [],
@@ -302,8 +302,8 @@ export default class BigCalendar extends Component {
 			mode === 'user' && view !== 'week' && newView === 'week'
 
 		await this.setState({
-			view: newView,
-			renderFirstCalendar: !movingToWeek
+			view: newView
+			// renderFirstCalendar: !movingToWeek
 		})
 
 		// because month view does not show all teammates, if we are in team mode jumping OFF month view, lets
@@ -311,10 +311,11 @@ export default class BigCalendar extends Component {
 		if (mode === 'team' && view === 'month' && newView !== 'month') {
 			this.toggleShowOnCalendars()
 		} else if (mode === 'user' && view !== 'week' && newView === 'week') {
+			//NOTE: Removed this delay as it was causing DOM issues with the calendar not rendering fast enough;
+			// Changes to BE data structure and FE should limit render lag that was initially seen
 			// week view is heavy, give dom a sec to render before rendering calendar
-			this.delayedRenderWeekView()
+			// this.delayedRenderWeekView()
 		}
-
 		this.handleChange()
 		//trigger a refresh which causes, sizes to be recalculated. 500 delay for css transitions
 		setTimeout(() => {
@@ -392,7 +393,10 @@ export default class BigCalendar extends Component {
 				.seconds(0)
 		}
 
-		return [earliest, latest]
+		return [
+			earliest.format('YYYY-MM-DD HH:mm:ss'),
+			latest.format('YYYY-MM-DD HH:mm:ss')
+		]
 	}
 
 	toggleShowOnCalendars = () => {
@@ -646,12 +650,14 @@ export default class BigCalendar extends Component {
 			view: selectedView,
 			formats,
 			toolbar: false,
-			date: selectedDate.toDate(),
-			min: min.toDate(),
-			max: max.toDate()
+			min: min,
+			max: max
 		}
 
 		// Determine selected date in relation to today
+		const currentDate = moment
+			.tz(selectedDate, auth.Location.timezone)
+			.format('YYYY-MM-DD HH:mm:ss')
 		const today = moment()
 			.tz(auth.Location.timezone)
 			.startOf('day')
@@ -695,7 +701,7 @@ export default class BigCalendar extends Component {
 			<div className={`big_calendar ${classNames}`}>
 				{isSelectingScheduleDate && (
 					<Dialog
-						title={`Jump To Day`}
+						title={`Choose Date`}
 						className={`schedule_calendar_select`}
 						onTapClose={this.handleHideScheduleDateDialog}
 					>
@@ -703,6 +709,7 @@ export default class BigCalendar extends Component {
 							defaultDate={selectedDate}
 							initialVisibleMonth={() => selectedDate}
 							onDateSelect={this.handleScheduleDateSelect}
+							timezone={auth.Location.timezone}
 							allowPastDates
 						/>
 						{!isToday && (
@@ -785,6 +792,7 @@ export default class BigCalendar extends Component {
 											className={`${
 												idx === 0 && !renderFirstCalendar ? 'hide' : ''
 											}`}
+											currentDate={currentDate}
 											views={views}
 											events={events ? this.filterEvents(events, teammate) : []}
 											eventPropGetter={event => this.applyClassNames(event)}