+++ packages/react-sprucebot/src/components/Calendar/Calendar.js
@@ -15,7 +15,15 @@ BigCalendar.setLocalizer(BigCalendar.momentLocalizer(moment)) // or globalizeLoc
 const CalendarComponent = withDragAndDrop(BigCalendar)
 
 class Calendar extends Component {
-	state = {}
+	state = {
+		today: null
+	}
+
+	componentDidMount() {
+		const { currentDate } = this.props
+
+		this.setState({ today: moment(currentDate).toDate() })
+	}
 
 	onNavigate = e => {
 		// Not fired with current build but causes error if omitted
@@ -31,12 +39,12 @@ class Calendar extends Component {
 	}
 
 	render() {
-		const {
-			defaultDate = new Date(),
-			canDrag,
-			canResize,
-			...props
-		} = this.props
+		const { today } = this.state
+		const { currentDate, canDrag, canResize, min, max, ...props } = this.props
+
+		const formattedDate = moment(currentDate).toDate()
+		const formattedMin = moment(min).toDate()
+		const formattedMax = moment(max).toDate()
 
 		return (
 			<div {...props} className={`calendar__wrapper`}>
@@ -46,8 +54,12 @@ class Calendar extends Component {
 					resizableAccessor={canResize}
 					startAccessor={this.startAccessor}
 					endAccessor={this.endAccessor}
-					defaultDate={defaultDate}
+					defaultDate={today}
+					date={formattedDate}
+					getNow={() => today}
 					selectable={props.onSelectSlot ? true : ''}
+					min={formattedMin}
+					max={formattedMax}
 					{...props}
 				/>
 			</div>